refactor: separate StoryOperations::Create into its own file
